Fix workflow when using maven-versions-plugin
The commits with the version bumps wouldn't get added to the
repository, leaving the repo in a bad state.
